initial_version

+++ python/arcticdb/storage_fixtures/azure.py
@@ -9,8 +9,12 @@ As of the Change Date specified in that file, in accordance with the Business So
 import re
 import shutil
 import uuid
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Optional, Union
 from tempfile import mkdtemp
+
+from arcticdb.util.utils import get_logger
+from tests.util.mark import LINUX
+
 
 from .api import *
 from .utils import get_ephemeral_port, GracefulProcessUtils, wait_for_server_to_come_up, safer_rmtree, get_ca_cert_for_testing
@@ -36,8 +40,17 @@ class AzureContainer(StorageFixture):
     client: Optional["ContainerClient"]
     _admin_client: Optional["ContainerClient"] = None
 
-    def _set_uri_and_client(self, auth: str):
-        from azure.storage.blob import ContainerClient, LinearRetry
+    def _get_policy(self) -> str:
+        from azure.storage.blob import LinearRetry
+        # The retry_policy instance will be modified by the pipeline, so cannot be constant
+        return {
+            "connection_timeout": 1, 
+            "read_timeout": 2, 
+            "retry_policy": LinearRetry(retry_total=3, backoff=1), 
+            "connection_verify": self.factory.client_cert_file}
+
+    def _set_uri_and_client_azurite(self, auth: str):
+        from azure.storage.blob import ContainerClient
 
         f = self.factory
         self.arctic_uri = (
@@ -46,21 +59,32 @@ class AzureContainer(StorageFixture):
         )
         # CA_cert_dir is skipped on purpose; It will be tested manually in other tests
 
-        # The retry_policy instance will be modified by the pipeline, so cannot be constant
-        policy = {"connection_timeout": 1, "read_timeout": 2, "retry_policy": LinearRetry(retry_total=3, backoff=1), "connection_verify": f.client_cert_file}
-        self.client = ContainerClient.from_connection_string(self.arctic_uri, self.container, **policy)
+        self.client = ContainerClient.from_connection_string(self.arctic_uri, self.container, **self._get_policy())
         # add connection_verify=False to bypass ssl checking
 
-    def __init__(self, factory: "AzuriteStorageFixtureFactory") -> None:
+
+    def __init__(self, factory: Union["AzuriteStorageFixtureFactory", "AzureStorageFixtureFactory"]) -> None:
+        from azure.storage.blob import ContainerClient
+
         super().__init__()
         self.factory = factory
-        self.container = f"container{str(uuid.uuid1())[:8]}"
-        self._set_uri_and_client(f"AccountName={factory.account_name};AccountKey={factory.account_key}")
+        if self.is_real_azure:
+            self.container = self.factory.default_container
+            self.arctic_uri = self.factory.get_arctic_uri()
+            self.client = ContainerClient.from_connection_string(self.arctic_uri, self.container, **self._get_policy())
+        else:
+            self.container = f"container{str(uuid.uuid1())[:8]}"
+            self._set_uri_and_client_azurite(f"AccountName={factory.account_name};AccountKey={factory.account_key}")
+            # __exit__() assumes this object owns the container, so always create and bail if exists:
+            self.client.create_container()
 
-        # __exit__() assumes this object owns the container, so always create and bail if exists:
-        self.client.create_container()
+    def is_real_azure(self) -> bool:
+        print("REAL AZURE")
+        return isinstance(self.factory, AzureStorageFixtureFactory)
 
     def _safe_enter(self):
+        from azure.storage.blob import ContainerClient
+
         if self.factory.enforcing_permissions:
             from azure.storage.blob import generate_container_sas
 
@@ -74,20 +98,37 @@ class AzureContainer(StorageFixture):
                 # https://learn.microsoft.com/en-us/rest/api/storageservices/create-service-sas#permissions-for-a-directory-container-or-blob
                 policy_id="main",
             )
-            self._set_uri_and_client("SharedAccessSignature=" + sas)
+            if self.is_real_azure():
+                # Things to do against real Azure
+                uri = f"{self.factory.get_arctic_uri()};SharedAccessSignature={sas}"
+                self.client = ContainerClient.from_connection_string(uri, self.container, **self._get_policy())
+            else:
+                # Container is in a mock env - Azurite
+                self._set_uri_and_client_azurite("SharedAccessSignature=" + sas)
 
     def __exit__(self, exc_type, exc_value, traceback):
+        if self.is_real_azure():
+            if self.factory.clean_bucket_on_fixture_exit:
+                self.factory.cleanup_container(self)
+                if len(self.libs_from_factory) > 0:
+                    get_logger().warning(f"Libraries not cleared remaining {self.libs_from_factory.keys()}")
+            
         if self.client:
-            if self._admin_client:
-                self._admin_client.delete_container(timeout=3)
-                self._admin_client.close()
-            else:
-                self.client.delete_container(timeout=3)
+            if not self.is_real_azure():
+                if self._admin_client:
+                    self._admin_client.delete_container(timeout=3)
+                    self._admin_client.close()
+                else:
+                    self.client.delete_container(timeout=3)
             self.client.close()
             self.client = None
 
     def create_test_cfg(self, lib_name: str) -> EnvironmentConfigsMap:
         cfg = EnvironmentConfigsMap()
+        if self.factory.default_prefix:
+            with_prefix = f"{self.factory.default_prefix}/{lib_name}"
+        else:
+            with_prefix = False
         add_azure_library_to_env(
             cfg=cfg,
             lib_name=lib_name,
@@ -95,7 +136,7 @@ class AzureContainer(StorageFixture):
             container_name=self.container,
             endpoint=self.arctic_uri,
             ca_cert_path=self.factory.client_cert_file,
-            with_prefix=False,  # to allow azure_store_factory reuse_name to work correctly
+            with_prefix=with_prefix,  # to allow azure_store_factory reuse_name to work correctly
         )
         return cfg
 
@@ -132,6 +173,8 @@ class AzuriteStorageFixtureFactory(StorageFixtureFactory):
     enforcing_permissions = False
     """Set to True to create AzureContainer with SAS authentication"""
 
+    clean_bucket_on_fixture_exit = True
+
     def __init__(self, port=0, working_dir: Optional[str] = None, use_ssl: bool = True, ssl_test_support: bool = True):
         self.http_protocol = "https" if use_ssl else "http"
         self.port = port or get_ephemeral_port(1)
@@ -166,3 +209,62 @@ class AzuriteStorageFixtureFactory(StorageFixtureFactory):
 
     def create_fixture(self) -> AzureContainer:
         return AzureContainer(self)
+
+    def cleanup_container(self, b: AzureContainer):
+        # Do nothing in this case
+        pass            
+
+
+class AzureStorageFixtureFactory(StorageFixtureFactory):
+
+    endpoint: str
+    account_name : str
+    account_key : str
+    connection_string : str = None
+    default_container: str = None
+    default_prefix: Optional[str] = None
+    client_cert_file : str = None
+    protocol: str = None
+    clean_bucket_on_fixture_exit = True
+
+    def __init__(self, native_config: Optional[dict] = None):
+        self.native_config = native_config
+        if LINUX:
+            AzureStorageFixtureFactory.client_cert_file = "/etc/ssl/certs/ca-certificates.crt"
+
+    def _safe_enter(self):
+        pass
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        pass
+
+    def __str__(self):
+        return f"[{type(self)}=Container:{self.default_container}], ConnectionString:{self.connection_string}"
+
+    def initialize_from_connection_sting(self, constr: str, container: str, prefix: str = None) -> None:
+        assert constr, "Azure connection string not available"
+        assert container, "Azure container not available"
+        AzureStorageFixtureFactory.connection_string = constr
+        AzureStorageFixtureFactory.account_name = re.search(r'AccountName=([^;]+)', constr).group(1)
+        AzureStorageFixtureFactory.account_key = re.search(r'AccountKey=([^;]+)', constr).group(1)
+        AzureStorageFixtureFactory.protocol = re.search(r'DefaultEndpointsProtocol=([^;]+)', constr).group(1)
+        endpoint_suffix = re.search(r'EndpointSuffix=([^;]+)', constr).group(1)
+        AzureStorageFixtureFactory.endpoint = f"{AzureStorageFixtureFactory.protocol}://{AzureStorageFixtureFactory.account_name}.blob.{endpoint_suffix}"
+        AzureStorageFixtureFactory.default_container = container
+        if prefix:
+            AzureStorageFixtureFactory.default_prefix = prefix
+
+    def get_arctic_uri(self):
+        url = f"azure://CA_cert_path={self.client_cert_file};Container={self.default_container};Path_prefix={self.default_prefix}"
+        if self.connection_string:
+            url += f"{url};{self.connection_string}"
+        else:
+            raise NotImplementedError("Constructing azure uri from current settings")
+        return url
+
+    def create_fixture(self) -> AzureContainer:
+        return AzureContainer(self)
+
+    def cleanup_container(self, b: AzureContainer):
+        b.slow_cleanup(failure_consequence="The following delete bucket call will also fail. ")            
+
